Split up policies and add tests
